requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
djangorestframework>=3.15
drf-spectacular>=0.27
gunicorn>=21.2
whitenoise>=6.6
python-dotenv>=1.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.1
django-cors-headers>=4.3
